log timings for each step in REPL
